requirements.txt
web3>=7.0.0
python-dotenv>=1.0.0
eth-utils>=2.1.0
hexbytes>=0.3.1